Feature/sprint 5

+++ light-life/src/Pages/Components/TargetHandler.js
@@ -1,8 +1,12 @@
 import React, { useEffect, useState } from "react";
+import { useLocation, useParams } from "react-router-dom";
 import firebase from "firebase/app";
 import "firebase/firestore";
-import { useLocation, useParams } from "react-router-dom";
+import Swal from "sweetalert2";
 import style from "../../style/target.module.scss";
+import image from "../../style/image.module.scss";
+import spinner from "../../images/loading.gif";
+import nothing from "../../images/nothing.svg";
 
 function TargetHandler({ target, setTarget }) {
   const db = firebase.firestore();
@@ -12,7 +16,12 @@ function TargetHandler({ target, setTarget }) {
   const [date, setDate] = useState({});
   const pathName = useLocation().pathname;
   const params = useParams();
-
+  const [display, setDisplay] = useState("inline-block");
+  useEffect(() => {
+    if (target) {
+      setDisplay("none");
+    }
+  });
   const getInputHandler = (e) => {
     const { name } = e.target;
     if (name === "startDate") {
@@ -60,41 +69,52 @@ function TargetHandler({ target, setTarget }) {
   };
 
   const bindRemoveTarget = (e) => {
-    setTargetIndex(e.target.id);
-
-    setTarget([...target.filter((t, index) => index !== +e.target.id)]);
-    db.collection("dietitians")
-      .doc(params.dID)
-      .collection("customers")
-      .doc(params.cID)
-      .collection("target")
-      .get()
-      .then((docs) => {
-        const docsArray = [];
-        docs.forEach((doc) => {
-          docsArray.push(doc.id);
-        });
-        const getID = docsArray.find((d, index) => index === +e.target.id);
-        return getID;
-      })
-      .then((res) => {
+    Swal.fire({
+      text: "確定刪除嗎?",
+      icon: "warning",
+      showCancelButton: true,
+      cancelButtonText: "取消",
+      confirmButtonText: "確定",
+      confirmButtonColor: "#1e4d4e",
+    }).then((res) => {
+      if (res.isConfirmed) {
+        setTargetIndex(e.target.id);
+        setTarget([...target.filter((t, index) => index !== +e.target.id)]);
         db.collection("dietitians")
           .doc(params.dID)
           .collection("customers")
           .doc(params.cID)
           .collection("target")
-          .doc(`${res}`)
-          .delete()
-          .then(() => {
-            console.log("delete");
+          .get()
+          .then((docs) => {
+            const docsArray = [];
+            docs.forEach((doc) => {
+              docsArray.push(doc.id);
+            });
+            const getID = docsArray.find((d, index) => index === +e.target.id);
+            return getID;
           })
-          .catch((error) => {
-            console.log("Error:", error);
+          .then((res) => {
+            db.collection("dietitians")
+              .doc(params.dID)
+              .collection("customers")
+              .doc(params.cID)
+              .collection("target")
+              .doc(`${res}`)
+              .delete()
+              .then(() => {
+                console.log("delete");
+              })
+              .catch((error) => {
+                console.log("Error:", error);
+              });
           });
-      });
+      }
+    });
   };
 
   const bindSaveHandler = (e) => {
+    console.log(e.target.id);
     db.collection("dietitians")
       .doc(params.dID)
       .collection("customers")
@@ -137,16 +157,23 @@ function TargetHandler({ target, setTarget }) {
         setIsEditing(false);
       });
   };
+  console.log(target);
+
   return (
     <>
-      {isEditing
-        ? target.map((t, index) =>
+      {target ? (
+        isEditing ? (
+          target.map((t, index) =>
             index === +targetIndex ? (
               <div key={index} className={style["customer-target"]}>
                 <div className={style["target-header"]}>
                   <div className={style["alter-button"]}>
                     <button onClick={bindSaveHandler} id={index}>
-                      儲存
+                      <i
+                        class="fa fa-floppy-o"
+                        aria-hidden="true"
+                        id={index}
+                      ></i>
                     </button>
                     {/* <i className="fa fa-trash-o" aria-hidden="true"></i> */}
                     {/* <i className="fa fa-trash" aria-hidden="true"></i> */}
@@ -235,14 +262,21 @@ function TargetHandler({ target, setTarget }) {
                 <div className={style["target-header"]}>
                   <div className={style["alter-button"]}>
                     <button onClick={bindEditHandler} id={index}>
-                      編輯
+                      <i
+                        class="fa fa-pencil"
+                        aria-hidden="true"
+                        id={index}
+                        onClick={bindEditHandler}
+                      ></i>
                     </button>
-                    <i
-                      id={index}
-                      className="fa fa-trash-o"
-                      aria-hidden="true"
-                      onClick={bindRemoveTarget}
-                    ></i>
+                    <button onClick={bindRemoveTarget} id={index}>
+                      <i
+                        id={index}
+                        className="fa fa-trash-o"
+                        aria-hidden="true"
+                        onClick={bindRemoveTarget}
+                      ></i>
+                    </button>
                   </div>
                   <div className={style.flexbox}>
                     <div className={style.title}>建立時間</div>
@@ -276,7 +310,8 @@ function TargetHandler({ target, setTarget }) {
               </div>
             )
           )
-        : target.map((t, index) =>
+        ) : target.length > 0 ? (
+          target.map((t, index) =>
             index === +targetIndex ? (
               <>
                 <div key={index} className={style["customer-target"]}>
@@ -284,14 +319,21 @@ function TargetHandler({ target, setTarget }) {
                     {pathName.includes("dietitian") ? (
                       <div className={style["alter-button"]}>
                         <button onClick={bindEditHandler} id={index}>
-                          編輯
+                          <i
+                            class="fa fa-pencil"
+                            aria-hidden="true"
+                            id={index}
+                            onClick={bindEditHandler}
+                          ></i>
                         </button>
-                        <i
-                          id={index}
-                          className="fa fa-trash-o"
-                          aria-hidden="true"
-                          onClick={bindRemoveTarget}
-                        ></i>
+                        <button onClick={bindRemoveTarget} id={index}>
+                          <i
+                            id={index}
+                            className="fa fa-trash-o"
+                            aria-hidden="true"
+                            onClick={bindRemoveTarget}
+                          ></i>
+                        </button>
                       </div>
                     ) : (
                       ""
@@ -343,14 +385,21 @@ function TargetHandler({ target, setTarget }) {
                   {pathName.includes("dietitian") ? (
                     <div className={style["alter-button"]}>
                       <button onClick={bindEditHandler} id={index}>
-                        編輯
+                        <i
+                          class="fa fa-pencil"
+                          aria-hidden="true"
+                          id={index}
+                          onClick={bindEditHandler}
+                        ></i>
                       </button>
-                      <i
-                        id={index}
-                        className="fa fa-trash-o"
-                        aria-hidden="true"
-                        onClick={bindRemoveTarget}
-                      ></i>
+                      <button onClick={bindRemoveTarget} id={index}>
+                        <i
+                          id={index}
+                          className="fa fa-trash-o"
+                          aria-hidden="true"
+                          onClick={bindRemoveTarget}
+                        ></i>
+                      </button>
                     </div>
                   ) : (
                     ""
@@ -386,7 +435,17 @@ function TargetHandler({ target, setTarget }) {
                 </div>
               </div>
             )
-          )}
+          )
+        ) : (
+          <div className={image.nothing}>
+            <img src={nothing} />
+          </div>
+        )
+      ) : (
+        <div className={image.spinner}>
+          <img src={spinner} style={{ display: display }} />
+        </div>
+      )}
     </>
   );
 }
